服务员结账-优惠券列表弹窗页面实现

+++ app/src/main/java/com/yonyou/hhtpos/adapter/ADA_WaitersList.java
@@ -17,7 +17,7 @@ import com.yonyou.hhtpos.bean.WaiterEntity;
  */
 public class ADA_WaitersList extends CommonAdapterListView<WaiterEntity> {
 
-    private double mSelectedPos;
+    private int mSelectedPos;
 
     public ADA_WaitersList(Context context) {
         super(context);
